requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.4
aiosignal==1.3.1
attrs==24.2.0
bitsandbytes==0.43.3
certifi==2024.7.4
charset-normalizer==3.3.2
datasets==2.21.0
dill==0.3.8
docstring_parser==0.16
einops==0.8.0
filelock==3.15.4
flash-attn==2.6.3
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.65.5
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
peft==0.12.0
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
shtab==1.7.1
six==1.16.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
trl==0.9.6
typing_extensions==4.12.2
tyro==0.8.8
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xxhash==3.5.0
yarl==1.9.4
